requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymongo
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-mongodb
python-dotenv
pandas
